Refactor code: added static functions to create Json objects.
